+++ src/ska_tmc_centralnode/manager/component_manager_mid.py
@@ -402,30 +402,28 @@ class CNComponentManagerMid(CNComponentManager):
         self._aggregate_state()
         self._update_imaging()
 
+    def get_dish_leaf_node_device_names(self) -> tuple:
+        """
+        Return Dish leaf node device names
+        """
+        return self.input_parameter.dish_leaf_node_dev_names
+
     def update_device_dish_mode(self, dev_name, dish_mode: DishMode) -> None:
         """
-        Update the dish mode of the given dish and call
-        the relative callbacks if available.
+        Update the dish mode of the given dish leaf node
+        and call the relative callbacks if available.
         :param dishMode: Dish mode of the device
         :type dishMode: DishMode
         """
-
         with self.lock:
             self.logger.info(
                 f"Dish event callback for device {dev_name}: {dish_mode}"
             )
 
-            # Update Dish Master device name with full FQDN for real Dish
-            dish_master_dev_names = self.get_dish_device_names()
-            self.logger.info(
-                "Dish master device names: %s", dish_master_dev_names
-            )
-            for dish in dish_master_dev_names:
-                # device name received in event is always in lower case for
-                # example. mid-dish/dish-manager/ska001 for dish manager 001.
-                # Therefore need to compare device name with dish name in
-                # lower case
-                if dev_name in dish.lower():
+            # Update Dish leaf node device name with full FQDN for real Dish
+            dish_leaf_node_dev_names = self.get_dish_leaf_node_device_names()
+            for dish in dish_leaf_node_dev_names:
+                if dev_name in dish:
                     dev_name = dish
 
             dev_info = self.component.get_device(dev_name)